+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-http/compare/1.0.1...HEAD)
 
+### Added
+
+- `HttpPanel` for Tracy - with request and response
+
 ## [1.0.1](https://github.com/orisai/nette-http/compare/1.0.0...1.0.1) - 2022-12-09
 
-- Composer
-	- allows PHP 8.2
-	- fix nette/http version constraint
+### Changed
+
+- Composer - allows PHP 8.2
+
+### Fixed
+
+- Composer - fix nette/http version constraint
 
 ## [1.0.0](https://github.com/orisai/nette-http/tag/1.0.0) - 2022-12-06